+++ app/src/main/java/com/nicoapps/cooktime/data/entity/InstructionEntity.kt
@@ -7,8 +7,8 @@ import java.time.Instant
 @Entity("instructions")
 data class InstructionEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val recipeId: Int = 0,
+    val id: Long = 0,
+    val recipeId: Long = 0,
     val text: String,
     val createdAt: Long = Instant.now().epochSecond,
     val lastUpdatedAt: Long? = null
